tested with double values

+++ src/cli/ExpEvalMain.java
@@ -5,7 +5,7 @@ import library.ExpEval;
 public class ExpEvalMain {
     public static void main(String[] args) {
         ExpEval expEval = new ExpEval();
-        int result = expEval.evaluator(args[0]);
+        int result = (int) expEval.evaluator(args[0]);
         System.out.println(result);
 
     }
